requirements.txt
aiohttp>=3.9
prompt_toolkit>=3.0
pyyaml>=6.0
